pyproject.toml
[tool.poetry]
name = "test-backend"
version = "0.1.0"
description = ""
authors = ["Daniialbek Daniialov"]

[tool.poetry.dependencies]
python = "^3.8"
django = "3.2"
djangorestframework = "^3.13.1"
psycopg2-binary = "^2.9.3"
djoser = "^2.1.0"
djangorestframework-simplejwt = "4.8.0"
drf-yasg = "^1.20.0"
django-extensions = "^3.1.5"
django-cors-headers = "^3.11.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
